requirements.txt
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.1
Django==5.1.5
idna==3.10
pillow==11.1.0
psycopg==3.2.4
requests==2.32.3
sqlparse==0.5.3
stripe==11.4.1
typing_extensions==4.12.2
urllib3==2.3.0
djangorestframework
rest_framework_simplejwt